Move InferConfig code for Gradle to separate file
